pyproject.toml
[project]
name = "matrix"
version = "0.1.0"
description = "全自动内容降维打击流水线"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "langgraph>=1.0.10",
    "jinja2>=3.1.0",
    "playwright>=1.40.0",
    "pillow>=12.1.1",
]

[project.scripts]
matrix = "matrix.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matrix"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
]
